# internal/platform/apidocs/swagger.yaml
definitions:
...
  apierr.Envelope:
    properties:
      error:
        $ref: '#/definitions/apierr.EnvelopeEr'
      request_id:
        example: req_example
        type: string
      type:
        enum:
        - error
        example: error
        type: string
    required:
    - error
    - request_id
    - type
    type: object
  apierr.EnvelopeEr:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        example: input is required
        type: string
      type:
        enum:
        - invalid_request_error
        - authentication_error
        - permission_error
        - not_found_error
        - request_too_large
        - rate_limit_error
        - api_error
        - timeout_error
        - overloaded_error
        example: invalid_request_error
        type: string
    required:
    - message
    - type
    type: object
...
  skills.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/skills.Skill'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
  skills.Skill:
    properties:
      created_at:
        format: date-time
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - created_at
    - description
    - id
    - name
    type: object
  vault.CreateRequest:
    properties:
      host:
        type: string
      name:
        type: string
      token:
        type: string
    required:
    - host
    - name
    - token
    type: object
  vault.Credential:
    properties:
      created_at:
        format: date-time
        type: string
      host:
        type: string
      id:
        type: string
      name:
        type: string
      revoked:
        type: boolean
    required:
    - created_at
    - host
    - id
    - name
    - revoked
    type: object
  vault.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/vault.Credential'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
info:
  contact: {}
  description: |-
    Managed Agents 服务。所有 /v1 接口使用 Bearer Wave API Key 认证； || Managed Agents service. All /v1 endpoints require a Bearer Wave API Key;
    由 `wave bootstrap` 签发。Authorization 填写完整的 `Bearer <Wave API Key>`。 || issue it with `wave bootstrap`. Set Authorization to the full `Bearer <Wave API Key>` value.
    offset 分页返回 data 和可选 next_offset（省略表示末页）；消息历史使用 next_after，事件列表使用最后一条 sequence 续读。 || Offset pagination returns data and an optional next_offset (omitted on the last page); message history uses next_after, and event lists resume from the last sequence.
    JSON 错误返回 {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}。文件传输的 Range 错误返回 text/plain；条件下载和健康检查可返回空响应。 || JSON errors return {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}. File-transfer Range errors return text/plain; conditional downloads and health checks may return empty responses.
    JSON 请求拒绝未知字段，非 multipart 请求体上限为 2 MiB。可空字段通过 x-nullable 标记；省略与 null 不等价。 || JSON requests reject unknown fields; non-multipart request bodies are limited to 2 MiB. Nullable fields use x-nullable; omission is not equivalent to null.
    交互式文档见 /swagger/index.html，OpenAPI spec 见 /swagger/doc.json。 || Interactive documentation: /swagger/index.html. OpenAPI spec: /swagger/doc.json.
  title: Wave AI API
  version: "1.0"
paths:
...
  /v1/credentials:
    get:
      description: 分页返回当前所有者可见的凭据元数据。凭据 token 只写不读，响应中永远不包含明文或密文。 || Return credential
        metadata visible to the current owner, with pagination. Tokens are write-only;
        responses never contain plaintext or ciphertext.
      operationId: vaultList
      parameters:
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vault.ListResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出凭据 || List credentials
      tags:
      - vault
    post:
      consumes:
      - application/json
      description: 为当前所有者录入一条主机凭据。凭据 token 仅在写入时加密存储，接口不提供任何读取明文的途径，响应中只返回元数据。 ||
        Store a host credential for the current owner. The token is encrypted on write;
        the API never exposes plaintext and returns metadata only.
      operationId: vaultCreate
      parameters:
      - description: 凭据名称、主机与 token || Credential name, host and token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/vault.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/vault.Credential'
        "400":
          description: 请求体不是合法 JSON 或 host 不是不含路径与用户信息的精确 authority || Invalid JSON
            body, or host is not an exact authority without a path or user information
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 录入凭据 || Store a credential
      tags:
      - vault
  /v1/credentials/{id}:
    delete:
      description: 将凭据标记为 revoked，此后不再可用于任务执行；凭据记录与密文保留用于审计，不提供明文读取。 || Mark the credential
        as revoked so it can no longer be used for task execution. Its record and
        ciphertext are retained for auditing; plaintext cannot be read.
      operationId: vaultRevoke
      parameters:
      - description: 凭据 ID || Credential ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 已吊销 || Revoked
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 凭据不存在或不在当前所有者范围内 || Credential not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 吊销凭据 || Revoke a credential
      tags:
      - vault
...
  /v1/skills:
    get:
      description: 分页返回当前所有者可见的技能包元数据，按主键排序。 || Return skill-package metadata visible
        to the current owner, paginated and ordered by primary key.
      operationId: skillsList
      parameters:
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/skills.ListResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出技能包 || List skill packages
      tags:
      - skills
    post:
      consumes:
      - multipart/form-data
      description: 通过 multipart/form-data 上传技能 ZIP 包（请求体上限 24 MiB）。ZIP 根目录必须包含带 YAML
        frontmatter（name 和 description）的 SKILL.md，最多 200 个文件且路径必须安全；包内容解压后上限 20 MiB。技能
        ID 不可变，创建后不可更换。 || Upload a skill ZIP using multipart/form-data, with a 24
        MiB request limit. The ZIP root must contain SKILL.md with YAML frontmatter
        (name and description). At most 200 files with safe paths are allowed; extracted
        content is limited to 20 MiB. Skill IDs are immutable and packages cannot
        be replaced after creation.
      operationId: skillsUpload
      parameters:
      - description: 包含根目录 SKILL.md 的 ZIP 技能包 || ZIP skill package containing SKILL.md
          at its root
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/skills.Skill'
        "400":
          description: 缺少 file 字段、ZIP 非法或超出大小/文件数限制、SKILL.md 缺失或元数据不合法 || Missing
            file field, invalid ZIP, exceeded size or file-count limits, missing SKILL.md
            or invalid metadata
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 上传技能包 || Upload a skill package
      tags:
      - skills
  /v1/skills/{id}/content:
    get:
      description: 按 ID 读取技能 ZIP 原包，经身份和归属校验后由服务端流式返回，文件名为 `<name>.zip`，按附件发送，支持 Range
        请求。 || Download the original skill ZIP by ID after authentication and ownership
        checks. The server streams it as an attachment named `<name>.zip`, with Range
        support.
      operationId: skillsContent
      parameters:
      - description: 技能 ID || Skill ID
        in: path
        name: id
        required: true
        type: string
      - description: 字节范围，如 bytes=0-1023 || Byte range, for example bytes=0-1023
        in: header
        name: Range
        type: string
      - description: 仅当 Last-Modified 日期匹配时返回范围，否则返回完整文件 || Return the range only
          if the Last-Modified date matches; otherwise return the full file
        in: header
        name: If-Range
        type: string
      - description: 未修改时返回 304（HTTP 日期） || Return 304 if unmodified (HTTP date)
        in: header
        name: If-Modified-Since
        type: string
      - description: 此日期后有修改时返回 412（HTTP 日期） || Return 412 if modified after this
          date (HTTP date)
        in: header
        name: If-Unmodified-Since
        type: string
      produces:
      - application/octet-stream
      - multipart/byteranges
      - application/json
      - text/plain
      responses:
        "200":
          description: 技能 ZIP 包二进制内容 || Skill ZIP package bytes
          headers:
            Accept-Ranges:
              description: bytes
              type: string
            Content-Disposition:
              description: attachment; filename=...
              type: string
            Last-Modified:
              description: HTTP 日期 || HTTP date
              type: string
          schema:
            type: file
        "206":
          description: 部分内容；多个范围返回 multipart/byteranges || Partial content; multiple
            ranges return multipart/byteranges
          headers:
            Accept-Ranges:
              description: bytes
              type: string
            Content-Disposition:
              description: attachment; filename=...
              type: string
            Content-Range:
              description: '单范围响应：bytes start-end/total || Single-range response:
                bytes start-end/total'
              type: string
            Last-Modified:
              description: HTTP 日期 || HTTP date
              type: string
          schema:
            type: file
        "304":
          description: 未修改，无响应体 || Not modified; empty body
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 技能不存在或不在当前所有者范围内 || Skill not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "412":
          description: 条件请求失败，无响应体 || Precondition failed; empty body
        "416":
          description: Range 无效或越界；底层下载错误为 text/plain || Invalid or unsatisfiable
            Range; underlying download errors use text/plain
          headers:
            Content-Range:
              description: '范围越界时：bytes */total || For an unsatisfiable range: bytes
                */total'
              type: string
          schema:
            type: string
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 下载技能包 || Download a skill package
      tags:
      - skills
...
securityDefinitions:
  BearerAuth:
    description: 填写完整的 Bearer <Wave API Key>，包含 Bearer 前缀和空格。 || Enter the full Bearer
      <Wave API Key> value, including the Bearer prefix and a space.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"